requirements.txt
streamlit==1.28.2
plotly==5.9.0
pandas==1.5.3
numpy==1.23.5
folium==0.14.0
streamlit-folium==0.16.0
Pillow==9.4.0
inflection==0.5.1